Fix double session read bug
Navigating to the sign in page throws an error if we have a loader on the _index route that reads the session. I don't know the reason for that so I decided to manually read the session on the sign in loader and pass it to the isAuthenticated method, which seems to solve the issue.